# swagger.yaml
openapi: 3.0.0
info:
  title: GeekMeet backend
  description: ""
  version: 0.0.1
servers:
  - url: /api/v2
...
paths:
...
  /friend/{id}:
    post:
      summary: Создать (отменить) заявку в друзья
      tags:
        - friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Действие произошло успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':    # status code
          description: Произошла ошибка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Убрать из друзей
      tags:
        - friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Удаление произошло успешно
        '400':    # status code
          description: Произошла ошибка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /friend/{id}/accept:
    patch:
      summary: Принять заявку в друзья
      tags:
        - friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Действие произошло успешно
        '400':    # status code
          description: Произошла ошибка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /friend/{id}/reject:
    patch:
      summary: Отклонить заявку в друзья
      tags:
        - friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Действие произошло успешно
        '400':    # status code
          description: Произошла ошибка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /friend/:
    get:
      summary: Получить все заявки вам в друзья
      tags:
        - friend
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Действие произошло успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    fromId:
                      type: integer
                    status:
                      type: string
        '400':    # status code
          description: Произошла ошибка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string